+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `liveion` WHIP push target (new `target-whip` feature): a configured stream target `url = "whip://[token@]host:port/whip/<stream>"` pushes the stream to a downstream WHIP endpoint — the declarative, static form of `cascade-push` (`POST /api/cascade/{stream}` with `target_url`). The push is media-driven: one supervisor task per target establishes the cascade-push session when the stream gains a publisher (`PublishStarted`, real WHIP or a source's virtual one) and tears it down when the publisher goes away, so the push always negotiates against the current publisher's codecs and downstream nodes see ordinary publisher attach/detach cycles (their `auto_delete_*`/on-demand strategies keep working). Failed pushes and lost sessions are retried with exponential backoff (5 s doubling, 60 s cap), event-bus lag is reconciled against the manager, and a target on an `on_demand` stream acts as standing demand that starts its sources at startup.
+- `liveion` WHIP push target (new `target-whip` feature): a configured stream target `url = "whip://[token@]host:port/whip/<stream>"` pushes the stream to a downstream WHIP endpoint — the declarative, static form of `cascade-push` (`POST /api/cascade/{stream}` with `target_url`). The push is media-driven: one supervisor task per target establishes the cascade-push session when the stream gains a publisher (`PublishStarted`, real WHIP or a source's virtual one) and tears it down when the publisher goes away, so the push always negotiates against the current publisher's codecs and downstream nodes see ordinary publisher attach/detach cycles (their `auto_delete_*`/on-demand strategies keep working). Failed pushes and lost sessions are retried with exponential backoff (5 s doubling, 60 s cap), event-bus lag is reconciled against the manager, and a target on an `on_demand` stream acts as standing demand: its sources are (re)started whenever the stream has neither a publisher nor a push session, paced by the same backoff, so the relay recovers on its own once an unreachable downstream is back.
 - `liveion` WHEP pull source (new `source-whep` feature, included in `source-all`): a configured stream source `url = "whep://[token@]host:port/whep/<stream>"` pulls media from an upstream WHEP endpoint — the declarative, static form of `cascade-pull` — and takes part in the full source lifecycle: `on_demand` start/stop, automatic reconnect, codec-readiness gating, and subscriber RTCP feedback (e.g. PLI keyframe requests) forwarded to the upstream publisher. The outgoing peer gathers ICE candidates with the server's own `[[ice_servers]]` config. livetwo exposes the shared `livetwo::whep::forward_rtcp_to_peer` helper used both here and by whepfrom's TCP output.
 - `liveion` on-demand sources: `[stream.<name>] on_demand = true` keeps the stream's configured sources (camera/encoder, RTSP pull, SDP) stopped until the first subscriber (WHEP, cascade push, or RTSP pull) arrives, and stops them again `on_demand_close_after_ms` after the last one leaves. The first subscriber waits up to `on_demand_start_timeout_ms` for the source to become ready so its SDP answer already contains the tracks; subscribers arriving while a start is in flight wait for it instead of receiving a track-less answer, and if the source is not ready in time the subscribe request fails (and cleans up) instead of hanging forever. The Dashboard shows `standby`/`on-demand` badges for these streams and `config` for other provisioned streams; `api::response::Stream` gained matching `provisioned`/`onDemand` fields.
 - `liveion` stream-lifecycle hooks: the new `[hooks]` section (global) and `[stream.<name>.hooks]` (per stream) run external scripts on stream lifecycle events — stream created/deleted (`on_stream_created`/`on_stream_deleted`) and publish started/stopped (`on_publish_started`/`on_publish_stopped`) — e.g. to power a capture device / hardware encoder on when media is actually wanted and off when the last consumer leaves. Publish hooks receive the publisher session id as `LIVE777_SESSION` (configured sources report `virtual-source`) and, for stop events, the reason (`peer-closed`/`api-deleted`/`idle-timeout`). Scripts receive `<event> <stream> [reason]` as argv and the same values as `LIVE777_EVENT`/`LIVE777_STREAM`/`LIVE777_REASON`; execution is a single FIFO queue (global hooks first, then per-stream, in configured order) with per-script `timeout_ms` and an `on_error = "stop"|"continue"` policy.
