+++ docs/zh/guide/live777.md
@@ -135,10 +135,11 @@ url = "whip://edge-1:7777/whip/cam1"
 推流是媒体驱动的：当流获得发布者（WHIP 发布端或配置的 source）时建立，
 发布者离开时拆除，因此下游节点看到的是普通的发布者上线/下线周期，
 下游自己的 `auto_delete_*`/on-demand 策略可以正常工作。在 `on_demand`
-流上配置的 target 相当于常驻需求，会在启动时拉起一次该流的源（之后
-如果流转为空闲，它会回到 standby 状态，而不会被反复拉起）。会话丢失
-后（下游关闭、ICE 失败、预注册流 reset）会以指数退避重建（5 秒起步
-翻倍，60 秒封顶）。
+流上配置的 target 相当于常驻需求：只要流既没有发布者也没有推流会话，
+就会（重新）拉起该流的源，因此下游从不可达恢复后继流会自动恢复。
+会话丢失后（下游关闭、ICE 失败、预注册流 reset）会以指数退避重建
+（5 秒起步翻倍，60 秒封顶）——on-demand 源的重新拉起也受同一退避约束，
+大约每分钟最多一次。
 
 ## DataChannel 转发
